+++ src/components/FeatureSelector/index.tsx
@@ -1,0 +1,3 @@
+export { default as AudioFeatureSlider } from './AudioFeatureSlider'
+export { default as FeatureMenu } from './FeatureMenu'
+export * from './FeatureSliderItem'
